requirements.txt
aiogram==3.15.0
telethon==1.37.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
openai==1.57.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2
cryptg==0.4.0
